Commit message: "Fix vector structure in ingestData function"
Description:
"Changed the structure of the vectors being ingested into Pinecone. Previously, the vectors field was a list containing a single item, which was another list of vectors. Now, the vectors field is a list of vectors. This fixes an error where Pinecone was expecting a vector dimension of 768 but was receiving a dimension of 0."